refactor: humble object pattern
