+++ app/src/main/java/by/profs/rowgame/data/competition/Ages.kt
@@ -12,7 +12,7 @@ private const val TY = 9
 private const val KID = 16
 private const val JUN = 18
 private const val YOUTH = 23
-private const val ADULT = 51
+private const val ADULT = 27
 private const val TY_C = 0.0
 private const val KID_C = 0.5
 private const val JUN_C = 0.7
