+++ app/src/main/java/by/profs/rowgame/data/consts/TableColumns.kt
@@ -10,5 +10,6 @@ const val NAME_ROWER = "name_surname"
 const val COL_AGE_CATEGORY = "age_category"
 const val COL_COMPETITION_LEVEL = "competition_level"
 const val COL_DAY = "day"
-const val COL_TYPE = "competition_type"
 const val COL_ROWER_AGE = "age"
+const val COL_STRATEGY = "strategy"
+const val COL_TYPE = "competition_type"
