+++ app/src/main/java/by/profs/rowgame/view/adapters/OarViewAdapter.kt
@@ -12,10 +12,11 @@ import androidx.fragment.app.findFragment
 import androidx.navigation.fragment.NavHostFragment
 import androidx.recyclerview.widget.RecyclerView
 import by.profs.rowgame.R
+import by.profs.rowgame.app.ServiceLocator
 import by.profs.rowgame.data.items.Oar
 import by.profs.rowgame.data.items.util.Manufacturer
 import by.profs.rowgame.data.preferences.PairingPreferences
-import by.profs.rowgame.presenter.database.MyRoomDatabase
+import by.profs.rowgame.presenter.database.dao.ComboDao
 import by.profs.rowgame.presenter.informators.OarInformator
 import by.profs.rowgame.presenter.informators.OarInformator.Companion.bladeImages
 import by.profs.rowgame.presenter.traders.OarTrader
@@ -30,14 +31,13 @@ import kotlinx.coroutines.launch
 class OarViewAdapter(
     private val oars: List<Oar>,
     private val type: Int,
-    infoBar: InfoBar,
-    private val database: MyRoomDatabase
+    infoBar: InfoBar
 ) : RecyclerView.Adapter<OarViewAdapter.ViewHolder>(), MyViewAdapter<Oar> {
 
     private lateinit var context: Context
     private lateinit var fragment: Fragment
     private val informator: OarInformator = OarInformator()
-    private val trader: OarTrader = OarTrader(infoBar, database.oarDao())
+    private val trader: OarTrader = OarTrader(infoBar)
     private val scope = CoroutineScope(Dispatchers.IO)
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
@@ -103,7 +103,7 @@ class OarViewAdapter(
                 val pairingPreferences = PairingPreferences(context)
                 pairingPreferences.occupyOar(oar.id!!)
                 scope.launch {
-                    database.comboDao().insertCombo(pairingPreferences.getCombo())
+                    ServiceLocator.get(ComboDao::class).insertCombo(pairingPreferences.getCombo())
                     val navController by lazy(LazyThreadSafetyMode.NONE) {
                         NavHostFragment.findNavController(fragment) }
                     MainScope().launch { PairingFragmentDirections.actionPairingFragmentSelf()
