+++ app/src/main/java/by/profs/rowgame/view/adapters/StandingViewAdapter.kt
@@ -8,6 +8,7 @@ import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import by.profs.rowgame.R
+import by.profs.rowgame.app.ServiceLocator
 import by.profs.rowgame.data.items.Rower
 import by.profs.rowgame.presenter.imageloader.loadThumb
 
@@ -17,7 +18,7 @@ class StandingViewAdapter(
     private val gap: ArrayList<Int> = ArrayList()
 ) : RecyclerView.Adapter<StandingViewAdapter.ViewHolder>() {
 
-    private lateinit var context: Context
+    private val context: Context = ServiceLocator.locate()
 
     class ViewHolder(view: View) : RecyclerView.ViewHolder(view) {
         val gap: TextView = view.findViewById(R.id.gap)
@@ -26,11 +27,8 @@ class StandingViewAdapter(
         val rowerPic: ImageView = view.findViewById(R.id.rower_pic)
     }
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-        context = parent.context
-        return ViewHolder(
-            LayoutInflater.from(context).inflate(R.layout.item_final_standing, parent, false))
-    }
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder = ViewHolder(
+        LayoutInflater.from(parent.context).inflate(R.layout.item_final_standing, parent, false))
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         if (mode != RESULTS) holder.gap.text = context.getString(
@@ -40,7 +38,7 @@ class StandingViewAdapter(
         val rower: Rower = standing[position]
         holder.position.text = (1 + position).toString()
         holder.name.text = rower.name
-        holder.rowerPic.loadThumb(rower)
+        holder.rowerPic.loadThumb(rower.thumb)
     }
 
     override fun getItemCount(): Int = standing.size
