+++ app/src/main/res/drawable/ic_training.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24"
+    android:tint="?attr/colorControlNormal">
+  <path
+      android:fillColor="@android:color/white"
+      android:pathData="M4.75,18.25c-0.41,0.41 -0.41,1.09 0,1.5 0.41,0.41 1.09,0.41 1.5,0L9,17h2l-2.5,-2.5 -3.75,3.75zM15,5c1.1,0 2,-0.9 2,-2s-0.9,-2 -2,-2 -2,0.9 -2,2 0.9,2 2,2zM20.29,20.3l-2,-2.01c-0.18,-0.18 -0.44,-0.29 -0.71,-0.29L16.5,18l-6.29,-6.29c0.79,-0.33 1.66,-0.87 2.29,-1.39v2.27l3.58,3.58c0.57,-0.55 0.92,-1.32 0.92,-2.16L17,8.26C17,7.02 15.98,6 14.74,6h-0.02c-0.34,0 -0.67,0.09 -0.96,0.23 -0.26,0.12 -0.5,0.29 -0.69,0.5l-1.4,1.55C10.61,9.45 8.66,10.35 7,10.32c-0.6,0 -1.08,0.48 -1.08,1.08 0,0.6 0.48,1.08 1.08,1.08 0.31,0 0.61,-0.03 0.9,-0.07l7.11,7.09v1.08c0,0.26 0.1,0.52 0.29,0.7l1.99,2.01c0.39,0.39 1.02,0.39 1.42,0l1.58,-1.58c0.39,-0.38 0.39,-1.02 0,-1.41z"/>
+</vector>
