+++ app/src/main/res/layout/fragment_inventory.xml
@@ -28,6 +28,7 @@
         android:layout_gravity="bottom|end"
         android:backgroundTint="@android:color/black"
         android:layout_margin="@dimen/fab_margin"
+        android:visibility="gone"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         android:src="@drawable/ic_add"
