+++ app/src/main/res/layout/item_pair.xml
@@ -19,7 +19,7 @@
         android:layout_height="@dimen/pair_picture_size"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toBottomOf="@id/rower_pic"
-        tools:src="@drawable/boat_single_scull" />
+        android:src="@drawable/boat_single_scull" />
 
     <ImageView
         android:id="@+id/boat_manuf_pic"
