--- app/src/main/res/layout/main_fragment.xml
@@ -1,61 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/main_fragment"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".view.activity.MainActivity">
-
-    <Button
-        android:id="@+id/go_to_boats"
-        style="@style/mainMenuButton"
-        android:background="@color/b1"
-        android:text="@string/boats"
-        app:layout_constraintBottom_toTopOf="@+id/go_to_oars"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_chainStyle="packed" />
-
-
-    <Button
-        android:id="@+id/go_to_oars"
-        style="@style/mainMenuButton"
-        android:background="@color/b2"
-        android:text="@string/oars"
-        app:layout_constraintBottom_toTopOf="@+id/go_to_rowers"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/go_to_boats" />
-
-    <Button
-        android:id="@+id/go_to_rowers"
-        style="@style/mainMenuButton"
-        android:background="@color/b3"
-        android:text="@string/rowers"
-        app:layout_constraintBottom_toTopOf="@id/go_to_new_pair"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/go_to_oars" />
-
-    <Button
-        android:id="@+id/go_to_new_pair"
-        style="@style/mainMenuButton"
-        android:background="@color/b5"
-        android:text="@string/new_pairs"
-        app:layout_constraintBottom_toTopOf="@+id/go_to_existing_pairs"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/go_to_rowers" />
-
-    <Button
-        android:id="@+id/go_to_existing_pairs"
-        style="@style/mainMenuButton"
-        android:background="@color/b6"
-        android:text="@string/trainings"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/go_to_new_pair" />
-</androidx.constraintlayout.widget.ConstraintLayout>
