+++ app/src/main/res/menu/bottom_nav.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/inventoryFragment"
+        android:icon="@drawable/ic_inventory"
+        android:title="@string/inventory" />
+
+    <item
+        android:id="@+id/shopFragment"
+        android:icon="@drawable/ic_shop"
+        android:title="@string/shop" />
+
+    <item
+        android:id="@+id/pairingFragment"
+        android:icon="@drawable/ic_pairing"
+        android:title="@string/new_pairs" />
+
+    <item
+        android:id="@+id/trainingFragment"
+        android:icon="@drawable/ic_training"
+        android:title="@string/trainings" />
+
+</menu>
