+++ app/src/main/res/navigation/nav_graph.xml
@@ -2,43 +2,17 @@
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/nav_graph"
-    app:startDestination="@id/mainFragment">
+    app:startDestination="@id/trainingFragment"
+    android:id="@+id/nav_graph">
 
-    <fragment
-        android:id="@+id/mainFragment"
-        android:name="by.profs.rowgame.view.fragments.MainFragment"
-        android:label="@string/app_name"
-        tools:layout="@layout/main_fragment">
-        <action
-            android:id="@+id/action_mainFragment_to_inventoryFragment"
-            app:destination="@id/inventoryFragment">
-            <argument
-                android:name="itemType"
-                android:defaultValue="6" />
-        </action>
-        <action
-            android:id="@+id/action_mainFragment_to_trainingFragment"
-            app:destination="@id/trainingFragment" />
-        <action
-            android:id="@+id/action_mainFragment_to_pairingFragment"
-            app:destination="@id/pairingFragment"
-            app:popUpTo="@id/mainFragment">
-            <argument
-                android:name="item"
-                android:defaultValue="1" />
-        </action>
-    </fragment>
     <fragment
         android:id="@+id/inventoryFragment"
         android:name="by.profs.rowgame.view.fragments.inventory.InventoryFragment"
+        android:label="@string/inventory"
         tools:layout="@layout/fragment_inventory">
         <argument
             android:name="itemType"
             app:argType="integer" />
-        <action
-            android:id="@+id/action_inventoryFragment_to_shopFragment"
-            app:destination="@id/shopFragment" />
         <action
             android:id="@+id/action_inventoryFragment_to_rowerDetailsFragment"
             app:destination="@id/rowerDetailsFragment" >
@@ -50,6 +24,7 @@
     <fragment
         android:id="@+id/shopFragment"
         android:name="by.profs.rowgame.view.fragments.shop.ShopFragment"
+        android:label="@string/shop"
         tools:layout="@layout/fragment_shop">
         <argument
             android:name="itemType"
@@ -87,8 +62,7 @@
         tools:layout="@layout/fragment_competition" >
         <action
             android:id="@+id/action_competitionFragment_to_trainingFragment"
-            app:destination="@id/trainingFragment"
-            app:popUpTo="@id/mainFragment" />
+            app:destination="@id/trainingFragment" />
     </fragment>
     <fragment
         android:id="@+id/pairingFragment"
@@ -101,8 +75,7 @@
             android:defaultValue="1" />
         <action
             android:id="@+id/action_pairingFragment_self"
-            app:destination="@id/pairingFragment"
-            app:popUpTo="@id/mainFragment">
+            app:destination="@id/pairingFragment">
             <argument
                 android:name="item"
                 android:defaultValue="1" />
@@ -128,7 +101,6 @@
         tools:layout="@layout/fragment_pre_competition">
         <action
             android:id="@+id/action_preCompetitionFragment_to_competitionFragment"
-            app:destination="@id/competitionFragment"
-            app:popUpTo="@id/mainFragment" />
+            app:destination="@id/competitionFragment" />
     </fragment>
 </navigation>
